fix bug

+++ mmaction/core/bbox/assigners/max_iou_assigner.py
@@ -129,7 +129,7 @@ class MaxIoUAssigner(BaseAssigner):
             # consider multi-class case (AVA)
             assert len(gt_labels[0]) > 1
             assigned_labels = assigned_gt_inds.new_zeros(
-                num_bboxes, len(gt_labels[0]))
+                num_bboxes, len(gt_labels[0]), dtype=torch.float32)
 
             # If not assigned, labels will be all 0
             pos_inds = torch.nonzero(assigned_gt_inds > 0).squeeze()
